Solucion de errores  varios

+++ models/Poyectos.js
@@ -5,7 +5,7 @@ const shortid = require('shortid');
 
 const Proyectos = db.define('proyectos', {
     id : {
-        type: DataTypes.INTEGER(11),
+        type: DataTypes.INTEGER(),
         primaryKey: true,
         autoIncrement: true
     },
